auth/iam: Add Service.IsBootstrapAdminAcccount
This commit adds a method to the IAM service to check whether a given
auth.Subject is the one true boostrap admin account.

It does so by verifying the UID and all permissions that the bootstrap admin
got assigned.

This is needed e.g. for the first start of the application. We need to show the
"Nutzerverwaltung" item in the menu, but we only want to show that
for an admin account. At this point in time no domain specific admin is created
yet and there was no way to canonically check whether the subject is the
bootstrap admin. Thus implementing logic to show/hide the item in the menubar
was not possibly to do correctly and safely

Signed-off-by: Frederick Lahde <frederick.lahde@worldiety.de>