feat: added scaffold footer support
The scaffold now supports a footer element. To support standardization, a small footer package with a simple default footer implementation is also provided. The ScaffoldBuilder can also be customized for a footer and provides an automatic footer generation based on the newly added corporate legal fields in the theme settings.