feat: supporting target when navigating forward
During tests, users in the registration process clicked on impress or the privacy policy and navigated inside the current tab as usual. However, when coming back, the forms were empty (which is how normal states work). The same happens on all interactive pages when clicking the according elements in the footer. Thus, any legal link from the footer and registration process opens now a new window or tab so that form contents are never lost.