feat: support to define arbitrary user consent options
More complex gdpr based consent use cases require that a user must be able to opt-in or opt-out for optional services and data exchange explicitly. The implementation deprecates the existing adoption flags in user.Settings but a migration path is implemented when using the predeclared consent identifiers. Furthermore, the entire history of the users approvals and revocations are tracked. Optional consent options are also displayed in the users profile menu and the user can freely toggle to approve or revoke a consent. Note, that there are is a different set of texts for the consent on the register page and the profile page, to allow a more legal description when registering.