feat!: removed existing inversion of control implementation and introducing a new context based api
The existing IoC implementation which is mainly used in the library code itself has an O(n) lookup time for services by name or type. It may also coerce in an unspecific way, if interface boxing looses type information. Also the configurator and application provide both different lookup implementations and in general this approach does not work with context propagation out of the box. This commit provides a O(1) implementation and provides a unified context based api. This is also suitable to customize data source for automatic form generation.