feat: TText can resolve its content against i18n translations
The crud package injects a lot of custom render views which are actually just TText values. However, at those places there are no subjects, windows or bundlers available, so that the actual string cannot get translated. The new resolve flag postpones the resolving until the TText is rendered and has a window with an i18n bundle available to translate the text.