feat: call window destroy observers on each invalidate not only if the window is actually destroyed
We noticed a lot of leak situations, especially if used with event bus observers, if registered as a window destroy observer. even though, the window is not destroyed but resetted on each invalidation (or render) cycle, this seems to conform to the expected behavior. The actual destruction of the window (or scope) may be still of interest, but we should introduce another kind of observer for that.