feat: added paths to scaffold builder which disables the default padding above the footer hline.
The default behavior is to provide a default padding, which simplifies UIs and seems to be the expected behavior. However, there are views (e.g. scrollviews which fill the scaffold body) which look wrong, if the padding is inserted because they must provide their own. So we keep the padding to not break existing layouts.