feat: added ai provider interface
The current architecture is to complex and error prone due to the decoupled local and cloud approach and the event-based sync mechanics. Instead, we now propose the provider interface, which is a lot simpler to implement for platforms and capability extensions. However, it makes various use cases slower and relies heavily on the availability of according cloud backend.