feat: added scroll to hash tag anchor support
To improve back and forth navigation, a simple support of intra-page-anchor support has been added. To trigger this, use the # as a key in core.Values and set it to the according id of your component. There are still a lot of edge cases, where the implementation cannot apply scrolling properly, however some simple cases already work.