feat: added label for property for Text
To improve accessiblity for checkbox and radiobutton, we need to connect the rendered input elements with a label using unique ids. To support this, we changed the behavior of the generated AutoState ids to be actually proper strings (which requires more allocations) but removing any headache with those IDs. Now a text can become a label for any other component referencing their id. TextField and the default picker.Picker renderer supports this behavior out of the box.